requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
colorama==0.4.6
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pdfkit==1.0.0
pyproject_hooks==1.0.0
PyYAML==6.0.1
requests==2.31.0
termcolor==2.3.0
urllib3==2.0.7